requirements.txt
Django==1.8.19
gunicorn==19.3.0
djangorestframework==3.1.1
django-bower==5.2.0
Pillow==2.8.1
Wand==0.4.4
django-imagekit==3.2.6
python3-memcached==1.51
drf-extensions==0.2.7
python-magic==0.4.15
